Refactor HTTP handler extension functions
The HTTP handlers have been refactored into extension functions for List<String>. The itemsHandler has been removed and replaced with a listHandler which is an extension function of List<String>. Therefore, the routes, now an extension of List<String>, have been refactored and now make use of listHandler. This change also affected the related tests where the items are now passed through the extension functions.