Add Todos class and update App.kt and AppTest.kt
A new Todos class was added in the kotlin files to enhance functionality and improve structuring. Due to this new class, App.kt and AppTest.kt were also modified to replace List with Todos. This allows for future growth and streamlining. Minor changes were made to the workspace.xml and other configurations.