Add deleteItemById function and corresponding test
A deleteItemById function has been added to the Todos class in Todos.kt. This provides a way to remove items from the Todos list by their unique ID. In addition, a new TodosTests.kt file has been created, which includes a test to verify the new functionality. The test checks that trying to delete an item not present in the list will be ignored.