Implement DELETE route and method for Todos
The DELETE HTTP method has been added to the application by implementing the `deleteHandler` and linked it to the "/item/{id}" route in the Todos class. Moreover, a test for this new feature has been provided in the AppTest file, ensuring that `deleteItemById` function properly removes an item from the Todos list.